+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,15 +1,12 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -20,7 +17,7 @@ public class UserService {
 
     private final UserStorage userStorage;
 
-    public User addFriend(Long idUser1, Long idUser2) {
+    public User addFriend(Integer idUser1, Integer idUser2) {
         if (idUser1 < 0 || idUser2 < 0 || idUser1 == null || idUser2 == null) {
             throw new ValidationException("Введено некорректное значение id");}
         User user1 = userStorage.getById(idUser1);
@@ -32,7 +29,7 @@ public class UserService {
         return user1;
     }
 
-    public User deleteFriend(Long idUser1, Long idUser2){
+    public User deleteFriend(Integer idUser1, Integer idUser2){
         if (idUser1 < 0 || idUser2 < 0 || idUser1 == null || idUser2 == null) {
             throw new ValidationException("Введено некорректное значение id");}
         User user1 = userStorage.getById(idUser1);
@@ -44,7 +41,7 @@ public class UserService {
         return user1;
     }
 
-    public List<User> getAllFriends(Long id) {
+    public List<User> getAllFriends(Integer id) {
         if (id < 0 || id == null) {
             throw new ValidationException("Введено некорректное значение id");}
        return userStorage.getById(id).getFriends().stream()
@@ -52,14 +49,14 @@ public class UserService {
                .collect(Collectors.toList());
     }
 
-    public List<User> getCommonFriends(Long idUser1, Long idUser2) {
+    public List<User> getCommonFriends(Integer idUser1, Integer idUser2) {
         if (idUser1 < 0 || idUser2 < 0 || idUser1 == null || idUser2 == null) {
             throw new NullPointerException("Введено некорректное значение id");
         }
         try {
             User user1 = userStorage.getById(idUser1);
             User user2 = userStorage.getById(idUser2);
-            Set<Long> commonFriendsId = user1.getFriends();
+            Set<Integer> commonFriendsId = user1.getFriends();
             commonFriendsId.retainAll(user2.getFriends());
             return commonFriendsId.stream()
                     .map(userStorage::getById)
